feat: 당첨된 로또의 개수와 당첨을 고려하여 수익률을 계산하는 기능 추가

+++ src/main/java/lotto/model/LottoResult.java
@@ -9,7 +9,7 @@ public class LottoResult {
     private int thirdCount;
     private int fourthCount;
     private int fifthCount;
-    private float profit;
+    private double profit;
 
     public void refreshResult(WinningRank winningRank) {
 
@@ -26,11 +26,11 @@ public class LottoResult {
         }
     }
 
-    public float getProfit() {
+    public double getProfit() {
         return profit;
     }
 
-    public void setProfit(float profit) {
+    public void setProfit(double profit) {
         this.profit = profit;
     }
